requirements.txt
lxml==6.0.0
python-docx==1.2.0
typing_extensions==4.14.1
flake8
black
isort
docx2pdf